[FIX]: Rename presenters.
TODO: Implement Timer.
