+++ onetodo/Presenter/TimePresenter.swift
@@ -1,0 +1,16 @@
+//
+//  TaskPresenter.swift
+//  onetodo
+//
+//  Created by Michel Ortega on 2/5/20.
+//  Copyright © 2020 Michel Ortega. All rights reserved.
+//
+
+import Foundation
+
+//	Protocol?
+
+class TimePresenter {
+	
+	
+}
